[#666] JUnit 5 Support
Refactored common code of XtextRunnerTest and InjectionExtensionTest
into an abstract base class.
