+++ org.eclipse.xtext.testing/tests/org/eclipse/xtext/testing/tests/InjectionExtensionTest.java
@@ -7,21 +7,9 @@
  *******************************************************************************/
 package org.eclipse.xtext.testing.tests;
 
-import org.eclipse.xtext.testing.IInjectorProvider;
-import org.eclipse.xtext.testing.IRegistryConfigurator;
 import org.eclipse.xtext.testing.InjectWith;
 import org.eclipse.xtext.testing.extensions.InjectionExtension;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import static org.junit.jupiter.api.Assertions.*;
-
-import com.google.inject.Binder;
-import com.google.inject.Guice;
-import com.google.inject.Inject;
-import com.google.inject.Injector;
-import com.google.inject.Module;
 
 /**
  * Test for {@link InjectionExtension}.
@@ -31,65 +19,5 @@ import com.google.inject.Module;
  */
 @InjectWith(InjectionExtensionTest.MyInjectorProvider.class)
 @ExtendWith(InjectionExtension.class)
-public class InjectionExtensionTest {
-
-	private static boolean injectorCreated = false;
-	private static boolean registrySaved = false;
-	private static boolean registryRestored = false;
-	
-	@Inject
-	private Boolean fieldsInjected = false;
-
-	public static class MyInjectorProvider implements IRegistryConfigurator, IInjectorProvider {
-
-		
-		@Override
-		public Injector getInjector() {
-			injectorCreated = true;
-			
-			assertTrue(registrySaved);
-			
-			return Guice.createInjector(new Module(){
-
-				@Override
-				public void configure(Binder binder) {
-					binder.bind(Boolean.class).toInstance(Boolean.TRUE);
-				}
-				
-			});
-		}
-
-		@Override
-		public void setupRegistry() {
-			registrySaved = true;
-			assertFalse(injectorCreated);
-		}
-
-		@Override
-		public void restoreRegistry() {
-			assertTrue(registrySaved);
-			registryRestored = true;
-		}
-		
-	}
-
-	
-	
-	@Before
-	public void beforeShouldBeExecutedAfterTheRegistriesAreInitialized(){
-		assertTrue(registrySaved);
-		assertTrue(injectorCreated);
-		assertTrue(fieldsInjected);
-	}
-	
-	@Test
-	public void shouldSaveRegistriesBeforeCreatingAnInjector() {
-		// tests are performed in MyInjectorProvider
-	}
-	
-	@After
-	public void afterShouldBeExecutedBeforeTheRegistriesAreRestored(){
-		assertFalse(registryRestored);
-	}
-
+public class InjectionExtensionTest extends AbstractJUnitIntegrationTest {
 }
